tests/wpt/mozilla/tests/mozilla/web_bluetooth.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var str = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
var filters = [];
for (var i = 0; i < str.length; ++i) {
    filters.push({ namePrefix: str.charAt(i)});
}

test(function() {
    window.testRunner.setBluetoothMockDataSet('NoNameHeartRateAdapter');

    var heart_rate_device = window.navigator.bluetooth.requestDevice(
        { filters: [{ services: ['generic_access', 'heart_rate'] }]}
    );
    assert_true(heart_rate_device instanceof BluetoothDevice);
    assert_equals(heart_rate_device.name, null);
    assert_true(heart_rate_device.gatt instanceof BluetoothRemoteGATTServer);
    assert_false(heart_rate_device.gatt.connected);
    var server = heart_rate_device.gatt.connect();
    assert_true(server instanceof BluetoothRemoteGATTServer);
    assert_throws('SecurityError', function () {
        heart_rate_device.gatt.getPrimaryService('battery_service');
    });
    assert_equals(server.device, heart_rate_device);

    var generic_access_service = server.getPrimaryService('generic_access');
    assert_true(generic_access_service instanceof BluetoothRemoteGATTService);
    assert_equals(generic_access_service.uuid, "00001800-0000-1000-8000-00805f9b34fb");
    var heart_rate_service = heart_rate_device.gatt.getPrimaryService('heart_rate');
    assert_true(heart_rate_device.gatt.getPrimaryService('heart_rate') instanceof BluetoothRemoteGATTService);
    assert_equals(heart_rate_service.uuid, "0000180d-0000-1000-8000-00805f9b34fb");

    var device_name_characteristic = generic_access_service.getCharacteristic(0x2a00);
    assert_true(device_name_characteristic instanceof BluetoothRemoteGATTCharacteristic);
    var pheripheral_privacy_flag_characteristic = generic_access_service.getCharacteristic(0x2a02);
    assert_true(pheripheral_privacy_flag_characteristic instanceof BluetoothRemoteGATTCharacteristic);
}, 'NoNameHeartRateAdapter Test');
</script>
